api/requirements.txt
Django==3.1.3
psycopg2-binary==2.8.6
python-dotenv==0.15.0
djangorestframework==3.12.2
djangorestframework-simplejwt==5.2.0